src/test/java/com/backspark/socksinventory/controller/SocksControllerImplTest.java
package com.backspark.socksinventory.controller;

import com.backspark.socksinventory.dto.SocksFilterRequest;
import com.backspark.socksinventory.dto.SocksRequest;
import com.backspark.socksinventory.dto.SocksResponse;
import com.backspark.socksinventory.exception.FileProcessingException;
import com.backspark.socksinventory.exception.NotEnoughSocksException;
import com.backspark.socksinventory.service.SocksService;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.web.multipart.MultipartFile;

import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.doThrow;
import static org.mockito.Mockito.when;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.multipart;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;

@WebMvcTest(SocksControllerImpl.class)
class SocksControllerImplTest {

    @Autowired
    private MockMvc mockMvc;

    @MockBean
    private SocksService socksService;

    @Test
    @DisplayName("GET /api/socks - should return socks count")
    void testGetSocksCount() throws Exception {
        // given
        when(socksService.getSocksCount(any(SocksFilterRequest.class))).thenReturn(100);

        // when-then
        mockMvc.perform(get("/api/socks")
                        .param("color", "red")
                        .param("operation", "moreThan")
                        .param("cottonPart", "50"))
                .andExpect(status().isOk())
                .andExpect(content().string("100"));
    }

    @Test
    @DisplayName("POST /api/socks/income - should add socks successfully")
    void testIncome() throws Exception {
        // given - нет необходимости что-то мока, метод void

        // language=JSON
        String requestJson = """
                {
                   "color": "red",
                   "cottonPart": 70,
                   "quantity": 10
                }
                """;

        // when-then
        mockMvc.perform(post("/api/socks/income")
                        .contentType(MediaType.APPLICATION_JSON)
                        .content(requestJson))
                .andExpect(status().isOk());
    }

    @Test
    @DisplayName("POST /api/socks/outcome - should return error if not enough socks")
    void testOutcomeNotEnough() throws Exception {
        // given
        doThrow(new NotEnoughSocksException("Not enough socks"))
                .when(socksService).removeSocks(any(SocksRequest.class));

        String requestJson = """
                {
                   "color": "red",
                   "cottonPart": 70,
                   "quantity": 100
                }
                """;

        // when-then
        mockMvc.perform(post("/api/socks/outcome")
                        .contentType(MediaType.APPLICATION_JSON)
                        .content(requestJson))
                .andExpect(status().isBadRequest())
                .andExpect(jsonPath("$.error").value("Not enough socks"))
                .andExpect(jsonPath("$.message").value("Not enough socks"));
    }

    @Test
    @DisplayName("PUT /api/socks/{id} - should update socks and return response")
    void testUpdateSocks() throws Exception {
        // given
        SocksResponse response = new SocksResponse();
        response.setId(1L);
        response.setColor("blue");
        response.setCottonPart(50);
        response.setQuantity(20);

        when(socksService.updateSocks(any(Long.class), any(SocksRequest.class)))
                .thenReturn(response);

        String requestJson = """
                {
                   "color": "blue",
                   "cottonPart": 50,
                   "quantity": 20
                }
                """;

        // when-then
        mockMvc.perform(put("/api/socks/1")
                        .contentType(MediaType.APPLICATION_JSON)
                        .content(requestJson))
                .andExpect(status().isOk())
                .andExpect(jsonPath("$.id").value(1L))
                .andExpect(jsonPath("$.color").value("blue"))
                .andExpect(jsonPath("$.cottonPart").value(50))
                .andExpect(jsonPath("$.quantity").value(20));
    }

    @Test
    @DisplayName("POST /api/socks/batch - should import socks batch")
    void testImportBatch() throws Exception {
        mockMvc.perform(multipart("/api/socks/batch")
                        .file("file", "red,70,10\nblue,50,20".getBytes()))
                .andExpect(status().isOk());
    }

    @Test
    @DisplayName("GET /api/socks - bad request if invalid parameters")
    void testGetSocksCountBadRequest() throws Exception {
        // Если validation или parse issue - например, передать невалидные параметры
        // Здесь можно проверить работу BindException или MethodArgumentNotValidException
        // Для упрощения просто не указываем cottonPart, хотя operation требует его.
        mockMvc.perform(get("/api/socks")
                        .param("color", "red")
                        .param("operation", "moreThan"))
                .andExpect(status().isOk());
        // В зависимости от логики валидации, можно ожидать 400,
        // но нужно быть увереным, что такая валидация есть в реальности.
        // Если нет валидации - этот тест может быть не актуален.
    }

    @Test
    @DisplayName("POST /api/socks/batch - FileProcessingException -> 422")
    void testFileProcessingException() throws Exception {
        doThrow(new FileProcessingException("Error processing file", null))
                .when(socksService).importSocksFromFile(any(MultipartFile.class));

        mockMvc.perform(multipart("/api/socks/batch")
                        .file("file", "invalid_line".getBytes()))
                .andExpect(status().isUnprocessableEntity())
                .andExpect(jsonPath("$.error").value("File processing error"))
                .andExpect(jsonPath("$.message").value("Error processing file"));
    }

    @Test
    @DisplayName("POST /api/socks/income - validation error -> 400")
    void testValidationError() throws Exception {
        String invalidRequestJson = """
                {
                   "color": "",
                   "cottonPart": 70,
                   "quantity": 10
                }
                """;

        mockMvc.perform(post("/api/socks/income")
                        .contentType(MediaType.APPLICATION_JSON)
                        .content(invalidRequestJson))
                .andExpect(status().isBadRequest())
                .andExpect(jsonPath("$.error").value("Validation error"));
    }

    @Test
    @DisplayName("GET /api/socks - generic exception -> 500")
    void testGenericException() throws Exception {
        // Смоделируем ситуацию, в которой сервис выбросит RuntimeException:
        when(socksService.getSocksCount(any())).thenThrow(new RuntimeException("Unexpected error"));

        mockMvc.perform(get("/api/socks")
                        .param("color", "red")
                        .param("operation", "equal")
                        .param("cottonPart", "50"))
                .andExpect(status().isInternalServerError())
                .andExpect(jsonPath("$.error").value("Internal server error"))
                .andExpect(jsonPath("$.message").value("Unexpected error"));
    }

}
